Improves color scheme for code snippet
