+++ src/styles.scss
@@ -147,7 +147,7 @@ body .p-component {
   max-width: 100%;
   border: .1rem solid rgba(0, 0, 0, .05);
   border-radius: .5rem;
-  background-color: #f4f4f4;
+  background-color: rgba(0, 0, 0, .05);
   text-align: left;
   overflow: scroll;
 
@@ -164,7 +164,11 @@ body .p-component {
     padding: 0 .2rem;
     border: .1rem solid rgba(0, 0, 0, .05);
     border-radius: .3rem;
-    background-color: #f4f4f4;
+    background-color: rgba(0, 0, 0, .05);
+  }
+
+  &.dark {
+    color: #fff;
   }
 }
